Teradata functions moved to com.facebook.presto.teradata.functions
