+++ presto-teradata-functions/pom.xml
@@ -48,7 +48,6 @@
         <dependency>
             <groupId>joda-time</groupId>
             <artifactId>joda-time</artifactId>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
