+++ presto-teradata-functions/src/main/java/com/facebook/presto/teradata/functions/TeradataDateFunctions.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teradata.presto.functions;
+package com.facebook.presto.teradata.functions;
 
 import com.facebook.presto.operator.Description;
 import com.facebook.presto.operator.scalar.ScalarFunction;
@@ -20,7 +20,7 @@ import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.type.StandardTypes;
 import com.facebook.presto.type.SqlType;
 import com.facebook.presto.util.ThreadLocalCache;
-import com.teradata.presto.functions.dateformat.TeradataDateFormatterBuilder;
+import com.facebook.presto.teradata.functions.dateformat.TeradataDateFormatterBuilder;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
 import org.joda.time.format.DateTimeFormatter;
