+++ presto-teradata-functions/src/main/java/com/facebook/presto/teradata/functions/dateformat/tokens/DDToken.java
@@ -11,33 +11,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teradata.presto.functions.dateformat.tokens;
+package com.facebook.presto.teradata.functions.dateformat.tokens;
 
 import com.facebook.presto.teradata.functions.DateFormatLexer;
-import com.teradata.presto.functions.dateformat.DateToken;
+import com.facebook.presto.teradata.functions.dateformat.DateToken;
 import org.joda.time.format.DateTimeFormatterBuilder;
 
 /**
- * Literal text token
+ * Day of month token
  */
-public class TextToken extends DateToken
+public class DDToken extends DateToken
 {
-    private String text;
-
-    public TextToken(String text)
-    {
-        this.text = text;
-    }
-
     @Override
     public Integer antlrToken()
     {
-        return DateFormatLexer.TEXT;
+        return DateFormatLexer.DD;
     }
 
     @Override
     public void appendTo(DateTimeFormatterBuilder builder)
     {
-        builder.appendLiteral(text);
+        builder.appendDayOfMonth(2);
     }
 }
