+++ presto-teradata-functions/src/main/java/com/facebook/presto/teradata/functions/dateformat/tokens/MIToken.java
@@ -11,26 +11,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.teradata.presto.functions.dateformat.tokens;
+package com.facebook.presto.teradata.functions.dateformat.tokens;
 
 import com.facebook.presto.teradata.functions.DateFormatLexer;
-import com.teradata.presto.functions.dateformat.DateToken;
+import com.facebook.presto.teradata.functions.dateformat.DateToken;
 import org.joda.time.format.DateTimeFormatterBuilder;
 
 /**
- * Day of month token
+ * Minute of hour token
  */
-public class DDToken extends DateToken
+public class MIToken extends DateToken
 {
     @Override
     public Integer antlrToken()
     {
-        return DateFormatLexer.DD;
+        return DateFormatLexer.MI;
     }
 
     @Override
     public void appendTo(DateTimeFormatterBuilder builder)
     {
-        builder.appendDayOfMonth(2);
+        builder.appendMinuteOfHour(2);
     }
 }
